backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scikit-learn==1.4.0
pytest==8.0.0
hypothesis==6.98.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy>=2.2.6
pandas==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
